backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
snowflake-connector-python==3.5.0
httpx==0.25.1
python-dotenv==1.0.0
itsdangerous==2.1.2
python-multipart==0.0.6

